@@ -17,19 +17,27 @@
   ],
   "license": "ISC",
   "dependencies": {
-    "@trinsic/okapi": "^1.0.0-preview.10",
-    "google-protobuf": "^3.17.0",
+    "@trinsic/okapi": "^1.0.0-preview.14",
+    "@trinsic/okapi-web": "^1.0.0-preview.14",
+    "google-protobuf": "^3.17.2",
     "grpc-web": "^1.2.1",
     "xmlhttprequest": "^1.8.0"
   },
   "devDependencies": {
     "@types/google-protobuf": "^3.15.2",
+    "@types/node": "^15.12.0",
     "ava": "^3.15.0",
     "copyfiles": "^2.4.1",
-    "grpc-tools": "^1.10.0",
-    "jest": "^26.6.3",
-    "ts-node": "^9.0.0",
-    "ts-protoc-gen": "^0.14.0",
-    "typescript": "^4.0.5"
+    "geckodriver": "^2.0.0",
+    "grpc-tools": "1.11.1",
+    "jest": "^27.0.4",
+    "ts-node": "^10.0.0",
+    "ts-protoc-gen": "0.15.0",
+    "typescript": "^4.3.2"
+  },
+  "ava": {
+    "require": [
+      "./test/_setup-browser-env.js"
+    ]
   }
 }
